+++ deploy/helpers/verifiers/passport.ts
@@ -18,6 +18,9 @@ import {
   PPerPassport_1_160_3_3_576_200_NAVerifier2__factory,
   PPerPassport_1_256_3_3_576_248_NAVerifier2__factory,
   PPerPassport_20_256_3_3_336_224_NAVerifier2__factory,
+  PPerPassport_10_256_3_3_576_248_1_1184_5_264Verifier2__factory,
+  PPerPassport_11_256_3_5_576_248_1_1808_4_256Verifier2__factory,
+  PPerPassport_21_256_3_3_576_232_NAVerifier2__factory,
   PUniversal2048Verifier2__factory,
   PUniversal2048V2Verifier2__factory,
   PUniversal4096Verifier2__factory,
@@ -48,6 +51,9 @@ export const deployPVerifiers = async (deployer: Deployer) => {
   await deployer.deploy(PPerPassport_1_160_3_3_576_200_NAVerifier2__factory);
   await deployer.deploy(PPerPassport_1_256_3_3_576_248_NAVerifier2__factory);
   await deployer.deploy(PPerPassport_20_256_3_3_336_224_NAVerifier2__factory);
+  await deployer.deploy(PPerPassport_10_256_3_3_576_248_1_1184_5_264Verifier2__factory);
+  await deployer.deploy(PPerPassport_11_256_3_5_576_248_1_1808_4_256Verifier2__factory);
+  await deployer.deploy(PPerPassport_21_256_3_3_576_232_NAVerifier2__factory);
 
   await deployer.deploy(PUniversal2048Verifier2__factory);
   await deployer.deploy(PUniversal2048V2Verifier2__factory);